refactor: getDate와 formatDate 함수를 date.js 파일로 통합
