+++ src/features/admin/approval/components/TableRow.jsx
@@ -3,7 +3,7 @@ import { db } from '@/shared/api/firebase/firebase'
 import { doc, getDoc, updateDoc } from 'firebase/firestore';
 import SelectBox from '@/shared/components/SelectBox/SelectBox';
 import Button from '@/shared/components/button/Button';
-import formatDate from '@/shared/utils/dateUtils'
+import { formatDate } from '@/shared/utils/date'
 import { Tr, ToggleImage } from '../TableCommonStyles'
 import { Td } from '../TableStyles';
