+++ src/features/admin/pay-stub/PayStub.jsx
@@ -3,7 +3,7 @@ import PayStubTable from './components/PayStubTable'
 import ContentWrap from '@/shared/components/content-wrap/ContentWrap'
 import Button from '@/shared/components/button/Button'
 import { useCallback, useEffect, useRef, useState } from 'react'
-import getDate from '../../../shared/utils/date'
+import { getDate } from '../../../shared/utils/date'
 import useThrottle from '../../../shared/hooks/useThrottle'
 import { useDispatch } from 'react-redux'
 import { upsertDocumentsForUsers } from './api/getCollectionWithFilter'
