+++ src/features/pay-stub/PayStub.jsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 import styled from 'styled-components'
-import getDate from '@/shared/utils/date'
+import { getDate } from '@/shared/utils/date'
 import ContentWrap from '@/shared/components/content-wrap/ContentWrap'
 import LoadingSpinner from '@/shared/components/loading-spinner/LoadingSpinner'
 import useFetchPayStub from './hooks/useFetchPayStub'
