+++ src/shared/utils/date.js
@@ -1,5 +1,5 @@
 // 공통 utils
-function getDate(part) {
+export function getDate(part) {
   const date = new Date()
 
   switch (part) {
@@ -14,4 +14,12 @@ function getDate(part) {
   }
 }
 
-export default getDate
+// ISO 8601 형식의 날짜 문자열을 'yyyy.mm.dd' 형식으로 변환
+export const formatDate = (isoString) => {
+  if (!isoString) return '';
+  const date = new Date(isoString);
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year}.${month}.${day}`;
+};
